Crate unic_ucd_version

UNIC — UCD — Core

A component of unic: Unicode and Internationalization Crates for Rust.

Core create indicating the version of Unicode Character Database.

Modules

pkg_info 🔒

Package information

unicode_version 🔒

Structs

UnicodeVersion

Represents a Unicode Version type.

Constants

PKG_DESCRIPTION

UNIC component description.

PKG_NAME

UNIC component name.

PKG_VERSION

UNIC component version.

UNICODE_VERSION

The Version of The Unicode Standard of the Unicode Character Database in use.